Move implementation files to src/
Also, move `convert.js` to `scripts/`.

This matches how our other libraries tend to be structured.
