+++ demo/src/distort/distort.js
@@ -5,8 +5,8 @@ const {
   createModelRenderer,
   createNode,
   setAttribute,
-} = require('../../../util');
-const foxJson = require('../../../fox.json');
+} = require('../../../src/util');
+const foxJson = require('../../../src/fox.json');
 
 createDistortedLogo({
   width: 0.4,
