+++ package.json
@@ -12,13 +12,13 @@
   },
   "license": "ISC",
   "author": "MetaMask",
-  "main": "index.js",
+  "main": "src/index.js",
   "files": [
-    "fox.json",
-    "util.js"
+    "src/fox.json",
+    "src/util.js"
   ],
   "scripts": {
-    "convert": "node convert.js",
+    "convert": "node scripts/convert.js",
     "demo": "node demo/build.js",
     "demo:all": "yarn demo --all",
     "lint": "yarn lint:eslint && yarn lint:misc --check",
