+++ scripts/convert.js
@@ -1,4 +1,5 @@
 const fs = require('fs').promises;
+const path = require('path');
 const { strict: assert } = require('assert');
 const OBJFile = require('obj-file-parser');
 const yargs = require('yargs/yargs');
@@ -101,19 +102,19 @@ async function main() {
     .usage('$0 [options]', usageDescription, (_yargs) =>
       _yargs
         .option('out', {
-          default: 'fox.json',
+          default: path.resolve(__dirname, '../src/fox.json'),
           description: 'The output file path',
           type: 'string',
           normalize: true,
         })
         .option('obj', {
-          default: 'fox.obj',
+          default: path.resolve(__dirname, '../src/fox.obj'),
           description: 'The input OBJ file path',
           type: 'string',
           normalize: true,
         })
         .option('mtl', {
-          default: 'fox.mtl',
+          default: path.resolve(__dirname, '../src/fox.mtl'),
           description: 'The input MTL file path',
           type: 'string',
           normalize: true,
