+++ src/fox.max
